Arrhenius.jl

<!– adapt from https://github.com/JuliaPOMDP/POMDPs.jl/blob/master/docs/src/index.md –> Combustion Kinetic Modeling Software 2.0.

Package Features

Arrhenius.jl is designed with following in mind:

Combustion software 2.0

Differential programing

Physics informed machine learning

Combustion simulation education.

Outline

Baiscs

Coming soon <!– ```@contents Pages = ["install.md", "get_started.md", "concepts.md"]

## Tutorials

<!-- ```@contents
Pages = [ "def_pomdp.md", "static.md", "interfaces.md", "dynamics.md"]

–>

Functions

Arrhenius.CreateSolution — Function

CreateSolution(mech)

Reaction mechanism is interepreted here. Part of the infomation are read in from the yaml file, pary of them are from the pre-processed .npz file from ReacTorch and Cantera test for math enviroment

Arrhenius.get_cp — Function

get specific of heat capacity

Arrhenius.get_cv — Function

get specific of heat capacity