API Documentation

Docstrings for Arrhenius.jl interface members can be accessed through Julia's built-in documentation system or in the list below.

You can use name(Arrhenius) to print a full list of exported namespace by Arrhenius.jl, and then copy the list below.

Arrhenius.C2X — Function

get mole fraction (X) from concentration (C)

Arrhenius.CreateSolution — Function

CreateSolution(mech)

Reaction mechanism is interepreted here. Part of the infomation are read in from the yaml file, pary of them are from the pre-processed .npz file from ReacTorch and Cantera test for math enviroment

Arrhenius.H_mass_func — Function

get enthaphy (H) per mass

Arrhenius.R — Constant

const R = 8314.46261815324 #J/(kmol*K)

Gas constant

Arrhenius.S_mass_func — Function

get entropy (S) per unit mass

Arrhenius.U_mass_func — Function

get enthaphy (U) per mass

Arrhenius.X2C — Function

get concentration (C) from mole fraction (X)

Arrhenius.X2Y — Function

get mass fraction (Y) from mole fraction (X)

Arrhenius.Y2C — Function

get concentration (C) from mass fraction (Y)

Arrhenius.Y2X — Function

get mole fraction (X) from mass fraction (Y)

Arrhenius.get_H — Function

get enthaphy (H) per mole

Arrhenius.get_S — Function

get entropy (S)

Arrhenius.get_U — Function

get enthaphy (U) per mole

Arrhenius.get_cp — Function

get specific of heat capacity

Arrhenius.get_cv — Function

get specific of heat capacity

Arrhenius.mix_trans — Function

mix_trans(gas::A, P, T, X, mean_MW) where {A <: Arrhenius.Solution}

Compute the tranposrt properties of a mixture using mixture average formula

Equations Ref. https://personal.ems.psu.edu/~radovic/ChemKinTheoryPaSR.pdf Equations. 5-50/51/52 for viscosity and thermal conductivity

Pure species viscosities [Pa-s]

Thermal conductivity. [W/m/K].

Equation 5-46 for diffusion

Mixture-averaged diffusion coefficients [m^2/s] relating the mass-averaged diffusive fluxes (with respect to the mass averaged velocity) to gradients in the species mole fractions.

Test this module in transporttest.jl

See also implementations in ReacTorch

Arrhenius.one_atm — Constant

const one_atm = 101325.0 #pascal

One atm

Arrhenius.set_states — Function

"mainly for testing code, will be removed in the future" "please customize such functions following this example"

Arrhenius.species_index — Function

Get species index of a species

Example

    species_index(gas, "O2")

Arrhenius.wdot_func — Function

compute reaction source term dC/dt

Thermo Interface API Documentation

Arrhenius.H_mass_func — Method

get enthaphy (H) per mass

Arrhenius.S_mass_func — Method

get entropy (S) per unit mass

Arrhenius.U_mass_func — Method

get enthaphy (U) per mass

Arrhenius.cal_a — Method

cal_a(gas, T, p, X)

calculates the molar helmholz free energy (a) for each species

Arrhenius.cal_a_RT — Method

cal_a_RT(gas, T, p, X)

calculates the dimensionless mole based helmholz free energy (a) for each species

Arrhenius.cal_a_mean — Method

cal_a_mean(gas, T, p, X)

calculates the mean mole based helmholz free energy (a) of the mixture

Arrhenius.cal_amass — Method

cal_amass(gas, T, p, X)

calculates the partial mass based helmholz free energy (a) for each species

Arrhenius.cal_amass_mean — Method

cal_amass_mean(gas, T, p, X)

calculates the mean mass based helmholz free energy (a) of the mixture

Arrhenius.cal_cp — Method

cal_cp(gas, T, p, X)

calculates the molar Heat capacity at constant pressure (cp) for each species

Arrhenius.cal_cp_R — Method

cal_cp_R(gas, T, p, X)

calculates the dimensionless mole based heat capacity at constant pressure (cp) for each species

Arrhenius.cal_cp_mean — Method

cal_cp_mean(gas, T, p, X)

calculates the mean mole based Heat capacity at constant pressure (cp) of the mixture

Arrhenius.cal_cpmass — Method

cal_cpmass(gas, T, p, X)

calculates the partial mass based Heat capacity at constant pressure (cp) for each species

Arrhenius.cal_cpmass_mean — Method

cal_cpmass_mean(gas, T, p, X)

calculates the mean mass based Heat capacity at constant pressure (cp) of the mixture

Arrhenius.cal_cv — Method

cal_cv(gas, T, p, X)

calculates the molar Heat capacity at constant volume (cv) for each species

Arrhenius.cal_cv_R — Method

cal_cv_R(gas, T, p, X)

calculates the dimensionless mole based heat capacity at constant volume (cv) for each species

Arrhenius.cal_cv_mean — Method

cal_cv_mean(gas, T, p, X)

calculates the mean mole based Heat capacity at constant volume (cv) of the mixture

Arrhenius.cal_cvmass — Method

cal_cvmass(gas, T, p, X)

calculates the partial mass based Heat capacity at constant volume (cv) for each species

Arrhenius.cal_cvmass_mean — Method

cal_cvmass_mean(gas, T, p, X)

calculates the mean mass based Heat capacity at constant volume (cv) of the mixture

Arrhenius.cal_g — Method

cal_g(gas, T, p, X)

calculates the molar gibbs free energy (g) for each species

Arrhenius.cal_g_RT — Method

cal_g_RT(gas, T, p, X)

calculates the dimensionless mole based free gibbs energy (g) for each species

Arrhenius.cal_g_mean — Method

cal_g_mean(gas, T, p, X)

calculates the mean mole based gibbs free energy (g) of the mixture

Arrhenius.cal_gmass — Method

cal_gmass(gas, T, p, X)

calculates the partial mass based gibbs free energy (g) for each species

Arrhenius.cal_gmass_mean — Method

cal_gmass_mean(gas, T, p, X)

calculates the mean mass based gibbs free energy (g) of the mixture

Arrhenius.cal_h — Method

cal_h(gas, T, p, X)

calculates the molar enthalpy (h) for each species

Arrhenius.cal_h_RT — Method

cal_h_RT(gas, T, p, X)

calculates the dimensionless mole based enthalpy (h) for each species

Arrhenius.cal_h_mean — Method

cal_h_mean(gas, T, p, X)

calculates the mean mole based enthalpy (h) of the mixture

Arrhenius.cal_hmass — Method

cal_hmass(gas, T, p, X)

calculates the partial mass based enthalpy (h) for each species

Arrhenius.cal_hmass_mean — Method

cal_hmass_mean(gas, T, p, X)

calculates the mean mass based enthalpy (h) of the mixture

Arrhenius.cal_s — Method

cal_s(gas, T, p, X)

calculates the molar entropy (s) for each species

Arrhenius.cal_s0 — Method

cal_s0(gas, T, p, X)

calculates the molar reference entropy (s0) for each species

Arrhenius.cal_s0_R — Method

cal_s0_R(gas, T, p, X)
cal_s0_R(gas, T)

calculates the dimensionless mole based reference state entropy (s0) for each species

Arrhenius.cal_s0_mean — Method

cal_s0_mean(gas, T, p, X)

calculates the mean mole based reference entropy (s0) of the mixture

Arrhenius.cal_s0mass — Method

cal_s0mass(gas, T, p, X)

calculates the partial mass based reference entropy (s0) for each species

Arrhenius.cal_s0mass_mean — Method

cal_s0mass_mean(gas, T, p, X)

calculates the mean mass based reference entropy (s0) of the mixture

Arrhenius.cal_s_R — Method

cal_s_R(gas, T, p, X)

calculates the dimensionless mole based entropy (s) for each species

Arrhenius.cal_s_mean — Method

cal_s_mean(gas, T, p, X)

calculates the mean mole based entropy (s) of the mixture

Arrhenius.cal_smass — Method

cal_smass(gas, T, p, X)

calculates the partial mass based entropy (s) for each species

Arrhenius.cal_smass_mean — Method

cal_smass_mean(gas, T, p, X)

calculates the mean mass based entropy (s) of the mixture

Arrhenius.cal_u — Method

cal_u(gas, T, p, X)

calculates the molar internal energy for each species

Arrhenius.cal_u_RT — Method

cal_u_RT(gas, T, p, X)

calculates the dimensionless mole based internal energy (u) for each species

Arrhenius.cal_u_mean — Method

cal_u_mean(gas, T, p, X)

calculates the mean mole based internal energy of the mixture

Arrhenius.cal_umass — Method

cal_umass(gas, T, p, X)

calculates the partial mass based internal energy for each species

Arrhenius.cal_umass_mean — Method

cal_umass_mean(gas, T, p, X)

calculates the mean mass based internal energy of the mixture

Arrhenius.get_H — Method

get enthaphy (H) per mole

Arrhenius.get_S — Method

get entropy (S)

Arrhenius.get_U — Method

get enthaphy (U) per mole

Arrhenius.get_cp — Method

get specific of heat capacity

Arrhenius.get_cv — Method

get specific of heat capacity